ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai>=0.5.0
python-dotenv==1.0.0
websockets==12.0
aiofiles==23.2.1
httpx==0.25.2
python-multipart==0.0.6 